+++ core/exports/effects.ts
@@ -10,5 +10,5 @@ export {
 } from "../src/effects/effects.ts";
 export { io, pluginIO } from "../src/effects/io.ts";
 export { config, denoConfig } from "../src/effects/config.ts";
-export { importmap, importmapHandlers } from "../src/effects/impormap.ts";
+export { importmap } from "../src/effects/impormap.ts";
 export { hot } from "../src/effects/hot-update.ts";
