+++ core/exports/plugins.ts
@@ -1,4 +1,5 @@
 export { pluginStripTypes } from "../src/plugins/strip-types.ts";
 export { pluginConfig } from "../src/plugins/config.ts";
+export { pluginImportmap } from "../src/plugins/importmap.ts";
 export { pluginRadish } from "../src/plugins/plugins.ts";
 export { pluginManifest, updateManifest } from "../src/plugins/manifest.ts";
